Revert "Revert "#16 Remove unused imports. Separate files""
This reverts commit dac90c2603429269efb6ebbfe25389a9c531bbdf.
